@dagar 2019-06-08
Here's the same comparison again, but at 1 kHz using the change from https://github.com/PX4/Firmware/pull/12145.

|| Relevant CPU (%) | Memory (bytes free) |
|----|----|----|
| PR #12145 (7e755b99449cd) | **18.82%** (fmu 8.96% + mc_att_control 9.87%) |39504 B|
| This PR @1 kHz (7f57a3be3cc) |**9.18%** (wq:rate_ctrl) |44560 B|