@dagar 2019-06-09
Splitting the usage out into separate PRs so we can focus on the core change.

1. px4fmu + mc_att_control: https://github.com/PX4/Firmware/pull/12223