@dagar 2019-06-10
Splitting the usage out into separate PRs so we can focus on the core change.

 - px4fmu: https://github.com/PX4/Firmware/pull/12224
 - mc_att_control: https://github.com/PX4/Firmware/pull/12225
 - fw_att_control: https://github.com/PX4/Firmware/pull/12226
 - ekf2: https://github.com/PX4/Firmware/pull/12227
 - fw_pos_control_l1: https://github.com/PX4/Firmware/pull/12228
 - vtol_att_control: https://github.com/PX4/Firmware/pull/12229