@dakejahl 2019-06-12
Looks awesome, this is definitely a better way of doing things. Just did a cursory of the code changes, I'm unfamiliar with the history of the work, but overall looks good.