@dagar 2019-06-13
> From a design point, what I'm still missing is how you plan to handle failover, and sensor timeouts in particular?

There's also a callback on `sensor_correction` that contains the selected gyro and only publishes when changed.

https://github.com/PX4/Firmware/pull/12225/commits/6516ae894d7c85114b6afefc82fb60b5ebcd3756#diff-69f93e040a677c931e797b5203c03387R101

https://github.com/PX4/Firmware/commit/6516ae894d7c85114b6afefc82fb60b5ebcd3756#diff-e1d4d75170cc48374d5f144ebfca43d9R105

As for the most other modules, they generally don't do anything on poll timeout other than setup the poll again. We'll still need to be careful and check things on a case-by-case basis.