@dagar 2019-06-13
Updated interval to microseconds to properly handle pwm rates (50-400 Hz or an interval of 2500 - 20000us).