@dagar 2019-06-14
The testing of this PR with several migrated modules (https://github.com/PX4/Firmware/pull/12256) looks good. Unless there's any objection let's merge this (not yet used anywhere) and continue iterating within the context of updating appropriate modules (https://github.com/PX4/Firmware/pull/12224, https://github.com/PX4/Firmware/pull/12225, https://github.com/PX4/Firmware/pull/12226, etc). If there are still concerns about failover or timeout handling it needs to be handled at that level anyway.