@dagar 2019-06-17
> Re timeout: that works then. I was under the impression you wanted to pull the rate controller into the sensor acquisition thread.

Still a possibility, but not if we don't need to. I'm actually considering going the other direction and pulling the filtering out of each driver and into the rate controller thread.